Restructuring the project
