+++ app/src/main/java/com/example/burgerapp/MainActivity.kt
@@ -12,7 +12,8 @@ import com.example.burgerapp.ui.theme.BurgerAppTheme
 import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
 import androidx.lifecycle.ViewModelProvider
-import com.google.android.gms.common.internal.StringResourceValueReader
+import com.example.burgerapp.viewmodel.AuthViewModel
+
 
 class MainActivity : ComponentActivity() {
