+++ app/src/main/java/com/example/burgerapp/viewmodel/AuthViewModel.kt
@@ -1,7 +1,8 @@
-package com.example.burgerapp
+package com.example.burgerapp.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.burgerapp.utils.AuthMessages
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.GoogleAuthProvider
